Struct pallet_plasm_validator::GenesisConfig

pub struct GenesisConfig<T: Config> {
    pub validators: Vec<T::AccountId>,
}

Genesis config for the module, allow to build genesis storage.

Fields
validators: Vec<T::AccountId>

Set of next era accounts that can validate blocks.

Implementations

impl<T: Config> GenesisConfig<T>

pub fn build_storage(&self) -> Result<Storage, String>

Build the storage for this module.

pub fn assimilate_storage(&self, storage: &mut Storage) -> Result<(), String>

Assimilate the storage for this module into pre-existing overlays.