Trait pallet_plasm_validator::Config

pub trait Config: Config {
    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
    type Time: Time;
    type RewardRemainder: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type Reward: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::PositiveImbalance>;
    type EraFinder: EraFinder<EraIndex, SessionIndex, MomentOf<Self>>;
    type ForSecurityEraReward: ForSecurityEraRewardFinder<BalanceOf<Self>>;
    type ComputeEraParam;
    type ComputeEra: ComputeEraOnModule<Self::ComputeEraParam>;
    type Event: From<Event<Self>> + Into<Self::Event>;
}

Associated Types

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>

The staking balance.

type Time: Time

Time used for computing era duration.

type RewardRemainder: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Tokens have been minted and are unused for validator-reward. Maybe, dapps-staking uses ().

type Reward: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::PositiveImbalance>

Handler for the unbalanced increment when rewarding a staker. Maybe, dapps-staking uses ().

type EraFinder: EraFinder<EraIndex, SessionIndex, MomentOf<Self>>

The information of era.

type ForSecurityEraReward: ForSecurityEraRewardFinder<BalanceOf<Self>>

The rewards for validators.

type ComputeEraParam

The return type of ComputeEraWithParam.

type ComputeEra: ComputeEraOnModule<Self::ComputeEraParam>

Acutually computing of ComputeEraWithParam.

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

Loading content...