Struct pallet_plasm_validator::Module

pub struct Module<T: Config>(_);

Implementations

impl<T: Config + 'static> Module<T>

pub fn untreated_era() -> EraIndex

The already untreated era is EraIndex.

pub fn elected_validators<K: EncodeLike<EraIndex>>(
key: K
) -> Option<Vec<T::AccountId>>

The currently elected validator set keyed by stash account ID.

pub fn validators() -> Vec<T::AccountId>

Set of next era accounts that can validate blocks.

impl<T: Config> Module<T>

pub fn reward_to_validators(
era: &EraIndex,
max_payout: &BalanceOf<T>
) -> BalanceOf<T>